test: rename functional to integration
