+++ tests/states/openconfig_bgp/data/integration_tests/full_config/expected_result_junos.txt
