+++ tests/states/openconfig_bgp/data/integration_tests/full_config/openconfig.json
