+++ tests/states/openconfig_bgp/data/integration_tests/v4_only/openconfig.json
