+++ tests/states/openconfig_bgp/data/integration_tests/v6_only/openconfig.json
