+++ tests/states/openconfig_bgp/data/integration_tests/with_extras/openconfig.json
